alter message type to int

+++ src/Domain/Api.Ai.Domain.DataTransferObject/Response/Message/CardMessageResponse.cs
@@ -42,7 +42,7 @@ namespace Api.Ai.Domain.DataTransferObject.Response.Message
 
         public override void SetMessageType()
         {
-            this.Type = Enum.Type.Text;
+            this.Type = (int)Enum.Type.Text;
         }
 
         #endregion
